add db_helpers module and util function for time diff
