+++ src/predictmod/utils.py
@@ -8,7 +8,6 @@ import decimal
 import uuid
 from bson.dbref import DBRef
 from bson.objectid import ObjectId
-from predictmod.app import mongo
 
 
 class MongoEncoder(json.JSONEncoder):
@@ -26,17 +25,6 @@ class MongoEncoder(json.JSONEncoder):
         if isinstance(value, decimal.Decimal):
             return str(value)
         return super(MongoEncoder, self).default(value, **kwargs)
-
-
-def get_active_model():
-    """Returns the id of the active model or None if there aren't any active models."""
-    query_result = mongo.db.active_model.find()
-    encoder = MongoEncoder()
-    docs = [json.loads(encoder.encode(doc)) for doc in query_result]
-    if len(docs) == 1:
-        return docs[0]['model_id']
-    else:
-        return None
 
 
 def add_timestamps(data, output, start_at, delta='1 week'):
@@ -89,3 +77,9 @@ def utcnow():
 
 def generate_uuid():
     return str(uuid.uuid1())
+
+
+def get_time_difference(t1, t2):
+    FMT = '%Y-%m-%dT%H:%M:%S'
+    delta = (datetime.datetime.strptime(t2, FMT) - datetime.datetime.strptime(t1, FMT))
+    return int(delta.total_seconds() / 60.0 / 60.0)
